GUI Usuarios (#50)
* Cambios respecto a usuarios y sus atributos y funciones

* Proyecto configurado para uso de databases


* Creado ejemplo de uso de MySQL

* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Cambios respecto a usuarios y sus atributos y funciones


* Arreglos en funciones del DAO, incluimos la clase NUsuario, cambio en los paquetes y la organización y empezamos la implementación de la ventana de registro de usuarios y limpieza de clases para concordar con rama main


* Creacion de la interfaz de registro de usuario y hecho los arreglos tras la revision


* Creacion de la interfaz de perfil de usuario y arreglo de conflictos con la lista de favoritos


* Creación de las principales GUI de usuario (Log in, Register, Forgot Password).


* Añadir clases de LogIn y ViewUtils a github, y función open de la GUILogIn.java


* Configuracion de algunos paneles de la interfaz de perfil de usuario


* Configuracion de algunos paneles de la interfaz de perfil de usuario


* Configuracion de algunos paneles de la interfaz de perfil de usuario


* Arreglo de las funciones de los botones de la interfaz de perfil de usuario y conexiones con la base de datos


* Arreglo de las funciones de los botones de la interfaz de perfil de usuario y conexiones con la base de datos


* Se añade las funciones para actualizar la suscripcion


* Se añade las funciones para actualizar la suscripcion


* Finalizada interfaz de perfil


* Borrar presentación de la rama de negocio/integración

* Arreglos en funciones del dao para guardar la informacion del usuario logueado


* Arreglos en funciones del dao para guardar la informacion del usuario logueado


* Arreglos en funciones del dao para guardar la informacion del usuario logueado


* Arreglo de funciones al unir con el resto


* Cambio de la interfaz de usuario a un CardPanel


* Arreglo de una funcion que fallaba en el BOUsuario


* arreglos para subir al main y que coincidan las cosas


* arreglos para subir al main y que coincidan las cosas


---------

+++ src/negocio/Usuario.java
@@ -14,7 +14,7 @@ public class Usuario {
     protected String pais;
     protected Suscripciones suscripcion;
     protected String direccion;
-    protected int saldo;
+    protected double saldo;
     protected int id;
     protected int id_cesta;
     protected List<Integer> favoritos;//tiene una lista de los id de los articulos favoritos
@@ -107,11 +107,11 @@ public class Usuario {
         this.direccion = direccion;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
